refactor: organize components MCP helpers
