+++ packages/tools/components-mcp/src/tools/searchAssets.ts
@@ -3,8 +3,8 @@ import { searchAssets } from '../utils/assetRegistry';
 import { formatAssetSummary } from '../utils/formatters';
 import { textResponse } from '../utils/response';
 
-import type { ComponentsMcpTool } from './toolTypes';
-import { getAssetKindArg, getNumberArg, getStringArg } from './toolUtils';
+import type { ComponentsMcpTool } from '../utils/toolTypes';
+import { getAssetKindArg, getNumberArg, getStringArg } from '../utils/toolUtils';
 
 function normalizeLimit(value: number | undefined): number {
   if (value === undefined) {
