+++ packages/tools/components-mcp/src/tools/searchColorTokens.ts
@@ -3,8 +3,8 @@ import { searchColorTokens } from '../utils/colorRegistry';
 import { formatColorTokenSummary } from '../utils/formatters';
 import { textResponse } from '../utils/response';
 
-import type { ComponentsMcpTool } from './toolTypes';
-import { getColorTokenCategoryArg, getNumberArg, getStringArg } from './toolUtils';
+import type { ComponentsMcpTool } from '../utils/toolTypes';
+import { getColorTokenCategoryArg, getNumberArg, getStringArg } from '../utils/toolUtils';
 
 function normalizeLimit(value: number | undefined): number {
   if (value === undefined) {
