+++ packages/tools/components-mcp/src/utils/toolTypes.ts
